pyproject.toml
[project]
name = "pypi-claimer"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
